refactor: apply changes to featured projects section in the Home page
